use JsonResult as results on url
