+++ keycloak_scanner/scanners/realm_scanner.py
@@ -1,25 +1,11 @@
 from typing import List
 
+from keycloak_scanner.scanners.json_result import JsonResult
 from keycloak_scanner.scanners.scanner import Scanner
 
 URL_PATTERN = '{}/auth/realms/{}'
 
-
-class Realm:
-
-    def __init__(self, name: str, url: str, json: dict):
-        self.name = name
-        self.url = url
-        self.json = json
-
-    def __repr__(self):
-        return f'<{self.name}, {self.url}, {self.json}>'
-
-    def __eq__(self, other):
-        if isinstance(other, Realm):
-            return self.name == other.name and self.url == other.url and self.json == other.json
-        return NotImplemented
-
+Realm = JsonResult
 
 Realms = List[Realm]
